pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "markdown>=3.9",
    "pillow>=11.3.0",
    "pygments>=2.19.2",
    "pymupdf>=1.26.4",
    "python-markdown>=0.1.0",
]
